Cleaned up code.

+++ soundhelix/src/com/soundhelix/util/ConsistentRandom.java
@@ -38,8 +38,8 @@ public class ConsistentRandom {
      */
     
     public int getInteger(int min,int max,long seed) {
-    	Random r = new Random((long)min*167852533l+(long)max*7531057l+constantSeed+seed);
-    	return min+r.nextInt(max-min);
+    	Random r = new Random(min * 167852533l + max * 7531057l + constantSeed + seed);
+    	return min + r.nextInt(max - min);
     }
 
     public int getInteger(int min,int max,Object seedObject) {
@@ -47,7 +47,7 @@ public class ConsistentRandom {
     }
 
     public boolean getBoolean(long seed) {
-    	Random r = new Random(millis+constantSeed+seed);
+    	Random r = new Random(millis + constantSeed + seed);
     	return r.nextBoolean();
     }
